feat: Add HASH_SALT environment variable to CD workflow
This commit adds the `HASH_SALT` environment variable to the Continuous Deployment (CD) workflow. The `HASH_SALT` value is stored securely in the GitHub secrets and is now injected into the CD workflow using the `echo` command. This addition ensures that the `HASH_SALT` is available during the CD process, enhancing the security of the application.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.