feat: Remove HASH_SALT environment variable from CD workflow
This commit removes the `HASH_SALT` environment variable from the Continuous Deployment (CD) workflow. The `HASH_SALT` value was previously injected into the CD workflow, but it is no longer needed and has been removed. This modification simplifies the workflow and reduces the exposure of sensitive information.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.