feat: Add changeUsername method to IAuthRepository
The code changes add a new method, `changeUsername`, to the `IAuthRepository` interface. This method allows users to change their username by providing their email, current username, new username, and password. The method returns a `ChangeUsernameReturnType` or `null` if the change is unsuccessful. This addition enhances the functionality of the authentication repository by providing a way for users to update their username.